refactor(team): extract invitation into team/invitation/ submodule
Mirrors the user/session/ pattern. Moves invitation_model.rs,
invitation_repository.rs, and invitation_surreal_repo.rs into a new
team/invitation/ subdirectory and adds invitation/service.rs (new
InvitationService<R, IR>) and invitation/rest.rs with all five handlers.

TeamService is simplified from TeamService<R, IR, TR> to TeamService<R, TR>;
all invitation methods and the assert_shared_team_admin helper move to
InvitationService. InvitationServiceHandle is wired as its own app_data in
main.rs and exposed via team/mod.rs. OpenAPI paths and test_helpers updated
accordingly.

Made-with: Cursor
