+++ backend/src/resources/team/invitation/mod.rs
@@ -1,0 +1,12 @@
+pub(crate) mod model;
+
+pub mod repository;
+pub use repository::TeamInvitationRepository;
+
+pub(crate) mod surreal_repo;
+pub use surreal_repo::SurrealTeamInvitationRepo;
+
+pub mod service;
+pub use service::{InvitationService, InvitationServiceHandle};
+
+pub mod rest;
